requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
stripe==9.12.0
shippo==3.9.0
python-dotenv==1.0.1
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0